refactor old stuff
